chore: move defaulter to pkg, so godownloader can use it
